requirements.txt
pytest
selene>=2.0.b16
allure-pytest
selenium